% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bwa_mem2}
\alias{bwa_mem2}
\title{Align the reads with bwa-mem2 and directly output bam file}
\usage{
bwa_mem2(metadata, input_dir, output_dir, reference, t = 2)
}
\arguments{
\item{metadata}{data frame;}

\item{input_dir}{character; path to the input directory}

\item{output_dir}{character; path to output directory}

\item{reference}{character; path to the fasta genome reference file. Bwa-mem2 created index must be in the same directory.}

\item{t}{numeric; number of threads to use; default-2}
}
\description{
Bwa-mem2 is faster than bwa-mem and produces the same results. This function also assigns Read group ID and sample name to the bam header,
which is necessary for downstream GATK4 analysis.
}
