% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fastqc}
\alias{fastqc}
\title{Perform fastqc on provided files}
\usage{
fastqc(metadata, input_dir, output_dir, t = 2)
}
\arguments{
\item{metadata}{data frame;}

\item{input_dir}{character; path to the input directory}

\item{output_dir}{character; path to output directory}

\item{t}{numeric; number of threads to use; default=2}
}
\description{
Perform fastqc on provided files
}
