% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MarkDuplicatesSpark}
\alias{MarkDuplicatesSpark}
\title{Mark duplicates with MarkDuplicatesSpark}
\usage{
MarkDuplicatesSpark(metadata, input_dir, output_dir, t = 2)
}
\arguments{
\item{metadata}{data frame;}

\item{input_dir}{character; path to the input directory}

\item{output_dir}{character; path to output directory}

\item{t}{numeric; number of threads to use; default-2}

\item{reference}{character; path to the fasta genome reference file. Bwa-mem2 created index must be in the same directory.}
}
\description{
MarkDuplicatesSpark is faster than standard MarkDuplicates from picard tools and produces same results
}
