% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Mutect2}
\alias{Mutect2}
\title{Detect variants with Mutect2}
\usage{
Mutect2(metadata, input_dir, output_dir, reference, parallel)
}
\arguments{
\item{metadata}{data frame;}

\item{input_dir}{character; path to the input directory}

\item{output_dir}{character; path to output directory}

\item{reference}{character; path to the fasta genome reference file. Bwa-mem2 created index must be in the same directory.}

\item{t}{numeric; number of threads to use; default-2}

\item{parallel;}{how many samples should be ran in parallel, each one will need around 15 GB of RAM}
}
\description{
Run Mutect2 with the option to run multiple samples in parallel.
}
