% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trimmomatic}
\alias{trimmomatic}
\title{Perform adapter trimming}
\usage{
trimmomatic(metadata, input_dir, output_dir, t = 2, min_length = 20)
}
\arguments{
\item{metadata}{data frame;}

\item{input_dir}{character; path to the input directory}

\item{output_dir}{character; path to output directory}

\item{t}{numeric; number of threads to use; default=2}

\item{min_length}{numeric; minimum read length to be kept; default=20}
}
\description{
Perform Illumina Universal adapter trimming and filter all reads under selected length (default 20).
}
